+++ examples/optic_solar_eclipse.py
@@ -5,7 +5,7 @@ from starplot import Moon, Binoculars, Observer
 from starplot.styles import PlotStyle, extensions
 
 # time of partial eclipse. total eclipse started at 15:13:46
-dt = datetime(2024, 4, 8, 14, 40, 47, 0, tzinfo=ZoneInfo("US/Eastern"))
+dt = datetime(2024, 4, 8, 14, 45, 0, 0, tzinfo=ZoneInfo("US/Eastern"))
 
 
 observer = Observer(
@@ -18,7 +18,7 @@ m = Moon.get(dt=observer.dt, lat=observer.lat, lon=observer.lon)
 
 op = m.create_optic(
     observer=observer,
-    optic=Binoculars(magnification=20, fov=65),
+    optic=Binoculars(magnification=30, fov=65),
     style=PlotStyle().extend(
         extensions.GRAYSCALE_DARK,
         extensions.OPTIC,