+++ src/starplot/models/constellation.py
@@ -16,7 +16,11 @@ class Constellation(CatalogObject, SkyObject):
     """
 
     boundary: Polygon | MultiPolygon
-    """Shapely Polygon of the constellation's boundary. Right ascension coordinates are in degrees (0...360)."""
+    """
+    Shapely Polygon of the constellation's boundary. Right ascension coordinates are in degrees (0...360).
+    
+    If the constellation's boundary crosses the 0-meridian, then this will be a MultiPolygon split at the meridian.
+    """
 
     star_hip_ids: list[int]
     """List of HIP ids for stars that are part of the _lines_ for this constellation."""
@@ -33,6 +37,13 @@ class Constellation(CatalogObject, SkyObject):
     
     **Important**: Starplot treats Serpens as two separate constellations to make them easier to work with programatically. 
     Serpens Caput has the `iau_id` of `ser1` and Serpens Cauda is `ser2`
+    """
+
+    border: LineString = None
+    """
+    Shapely LineString of the constellation's border. Right ascension coordinates are in degrees (0...360).
+    
+    Coordinates in this geometry may extend past RA=360, if the border crosses the 0-meridian. This makes it ready to plot with the `line` function.
     """
 
     def __repr__(self) -> str: