+++ src/starplot/plots/base.py
@@ -8,7 +8,7 @@ from matplotlib import pyplot as plt, patheffects
 from matplotlib.axes import Axes
 from matplotlib.figure import Figure
 from matplotlib.lines import Line2D
-from shapely import Polygon
+from shapely import Polygon, LineString
 
 from starplot.coordinates import CoordinateSystem
 from starplot import geod, models, warnings
@@ -334,9 +334,9 @@ class BasePlot(TextPlotterMixin, ABC):
                 polygon_style = style.marker.to_polygon_style()
                 polygon_style.edge_color = None
                 self.circle(
-                    (p.ra, p.dec),
-                    p.apparent_size,
-                    polygon_style,
+                    center=(p.ra, p.dec),
+                    radius_degrees=p.apparent_size / 2,
+                    style=polygon_style,
                     gid="planet-marker",
                 )
                 self._add_legend_handle_marker(legend_label, style.marker)
@@ -406,10 +406,11 @@ class BasePlot(TextPlotterMixin, ABC):
             polygon_style.edge_color = None
 
             self.circle(
-                (s.ra, s.dec),
-                s.apparent_size,
+                center=(s.ra, s.dec),
+                radius_degrees=s.apparent_size / 2,
                 style=polygon_style,
                 gid="sun-marker",
+                num_pts=200,
             )
 
             style.marker.symbol = MarkerSymbolEnum.CIRCLE
@@ -633,14 +634,27 @@ class BasePlot(TextPlotterMixin, ABC):
             )
 
     @use_style(LineStyle)
-    def line(self, coordinates: list[tuple[float, float]], style: LineStyle, **kwargs):
+    def line(
+        self,
+        style: LineStyle,
+        coordinates: list[tuple[float, float]] = None,
+        geometry: LineString = None,
+        **kwargs,
+    ):
         """Plots a line
 
         Args:
             coordinates: List of coordinates, e.g. `[(ra, dec), (ra, dec)]`
+            geometry: A shapely LineString. If this value is passed, then the `coordinates` kwarg will be ignored.
             style: Style of the line
         """
-        x, y = zip(*[self._prepare_coords(*p) for p in coordinates])
+
+        if coordinates is None and geometry is None:
+            raise ValueError("Must pass coordinates or geometry when plotting lines.")
+
+        coords = geometry.coords if geometry is not None else coordinates
+
+        x, y = zip(*[self._prepare_coords(*p) for p in coords])
 
         self.ax.plot(
             x,
@@ -700,9 +714,9 @@ class BasePlot(TextPlotterMixin, ABC):
 
             if show_phase:
                 self._moon_with_phase(
-                    m.phase_description,
-                    (m.ra, m.dec),
-                    m.apparent_size,
+                    moon_phase=m.phase_description,
+                    center=(m.ra, m.dec),
+                    radius_degrees=m.apparent_size / 2,
                     style=polygon_style,
                     dark_side_color=style.marker.edge_color,
                 )
@@ -797,8 +811,8 @@ class BasePlot(TextPlotterMixin, ABC):
         # Plot left side
         self.ellipse(
             center,
-            radius_degrees * 2,
-            radius_degrees * 2,
+            height_degrees=radius_degrees * 2,
+            width_degrees=radius_degrees * 2,
             style=left,
             num_pts=num_pts,
             angle=0,
@@ -808,8 +822,8 @@ class BasePlot(TextPlotterMixin, ABC):
         # Plot right side
         self.ellipse(
             center,
-            radius_degrees * 2,
-            radius_degrees * 2,
+            height_degrees=radius_degrees * 2,
+            width_degrees=radius_degrees * 2,
             style=right,
             num_pts=num_pts,
             angle=180,
@@ -819,8 +833,8 @@ class BasePlot(TextPlotterMixin, ABC):
         # Plot middle
         self.ellipse(
             center,
-            radius_degrees * 2,
-            radius_degrees,
+            height_degrees=radius_degrees * 2,
+            width_degrees=radius_degrees,
             style=middle,
             gid="moon-marker",
         )